pdn/pyproject.toml
[project]
name = "persistent-diamonds-v3"
version = "0.1.0"
description = "Persistent Diamonds v3: IQT-informed architecture research package"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Diwank Singh Tomer" }]
dependencies = [
  "numpy>=1.26",
  "torch>=2.3",
  "transformers>=4.44",
  "typer>=0.12",
  "pyyaml>=6.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
training = [
  "datasets>=2.21",
  "accelerate>=0.34",
]
flash = [
  "flash-attn>=2.6; platform_system == 'Linux'",
]
dev = [
  "pytest>=8.3",
  "ruff>=0.6",
]

[project.scripts]
pdv3 = "persistent_diamonds_v3.cli:app"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/persistent_diamonds_v3"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
